pyproject.toml
[project]
name = "gemini-blog-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.8.0",
    "sqlalchemy>=2.0.40",
    "streamlit>=1.44.0",
]
